requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1

# AI & ML
openai-whisper
google-generativeai==0.8.3

# YouTube
yt-dlp==2024.12.13

# Utilities
Pillow
